New file for histogram view
